+++ regression/run/clibrary/clibrary.h
@@ -50,6 +50,8 @@ bool ImpliedBoolFalse(bool flag);
 void bindC1(void);
 void bindC2(const char * name);
 
+void passVoidStarStar(void *in, void **out);
+
 int passStruct1(Cstruct1 *s1);
 int passStruct2(Cstruct1 *s1, const char *name);
 Cstruct1 *returnStructPtr1(int ifield);